@@ -21,18 +21,18 @@
   "private": true,
   "dependencies": {
     "@angular-mdc/web": "^5.1.1",
-    "@angular/animations": "~12.2.16",
-    "@angular/cdk": "~12.2.13",
-    "@angular/common": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/flex-layout": "~12.0.0-beta.35",
-    "@angular/forms": "~12.2.16",
-    "@angular/material": "~12.2.13",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/pwa": "~12.2.17",
-    "@angular/router": "~12.2.16",
-    "@angular/service-worker": "~12.2.16",
+    "@angular/animations": "~13.3.11",
+    "@angular/cdk": "~13.3.9",
+    "@angular/common": "~13.3.11",
+    "@angular/core": "~13.3.11",
+    "@angular/flex-layout": "^13.0.0-beta.38",
+    "@angular/forms": "~13.3.11",
+    "@angular/material": "~13.3.9",
+    "@angular/platform-browser": "~13.3.11",
+    "@angular/platform-browser-dynamic": "~13.3.11",
+    "@angular/pwa": "^13.3.9",
+    "@angular/router": "~13.3.11",
+    "@angular/service-worker": "~13.3.11",
     "@auth0/auth0-spa-js": "^1.21.0",
     "@bugsnag/js": "^7.16.5",
     "@bugsnag/plugin-angular": "^7.16.5",
@@ -52,7 +52,7 @@
     "lodash-es": "^4.17.21",
     "mingo": "^4.4.1",
     "ngx-avatar": "^4.1.0",
-    "ngx-cookie-service": "~12.0.3",
+    "ngx-cookie-service": "^13.2.1",
     "ngx-quill": "14.1.1",
     "ot-json0": "^1.1.0",
     "papaparse": "^5.3.2",
@@ -71,16 +71,16 @@
     "zone.js": "^0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.17",
-    "@angular-eslint/builder": "^12.7.0",
-    "@angular-eslint/eslint-plugin": "^12.7.0",
-    "@angular-eslint/eslint-plugin-template": "^12.7.0",
-    "@angular-eslint/schematics": "12.7.0",
-    "@angular-eslint/template-parser": "^12.7.0",
-    "@angular/cli": "~12.2.17",
-    "@angular/compiler": "~12.2.16",
-    "@angular/compiler-cli": "~12.2.16",
-    "@angular/language-service": "~12.2.16",
+    "@angular-devkit/build-angular": "~13.3.9",
+    "@angular-eslint/builder": "^13.5.0",
+    "@angular-eslint/eslint-plugin": "^13.5.0",
+    "@angular-eslint/eslint-plugin-template": "^13.5.0",
+    "@angular-eslint/schematics": "^13.5.0",
+    "@angular-eslint/template-parser": "^13.5.0",
+    "@angular/cli": "~13.3.9",
+    "@angular/compiler": "~13.3.11",
+    "@angular/compiler-cli": "~13.3.11",
+    "@angular/language-service": "~13.3.11",
     "@triviality/logger": "^1.1.6",
     "@types/auth0-js": "^9.14.6",
     "@types/events": "^3.0.0",
@@ -115,7 +115,7 @@
     "prettier": "^2.6.2",
     "ts-mockito": "^2.6.1",
     "ts-node": "^10.7.0",
-    "typescript": "~4.3.5",
+    "typescript": "~4.6.4",
     "webpack-bundle-analyzer": "^4.5.0"
   },
   "browser": {
